% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb-public-api.R
\name{cran_events}
\alias{cran_events}
\title{List of all CRAN events (new, updated, archived packages)}
\usage{
cran_events(releases = TRUE, archivals = TRUE, limit = 10, from = 1)
}
\arguments{
\item{releases}{Whether to include package releases.}

\item{archivals}{Whether to include package archivals.}

\item{limit}{Number of events to list.}

\item{from}{Where to start the list, for pagination.}
}
\value{
List of events.
}
\description{
List of all CRAN events (new, updated, archived packages)
}
\examples{
\dontrun{
cran_events()
cran_events(limit = 5, releases = FALSE)
cran_events(limit = 5, archivals = FALSE)
}
}
