% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb-public-api.R
\name{cran_package}
\alias{cran_package}
\title{Metadata about a CRAN package}
\usage{
cran_package(name, version = NULL)
}
\arguments{
\item{name}{Name of the package.}

\item{version}{The package version to query. If \code{NULL}, the latest
version if returned.}
}
\value{
The package metadata, in a named list.
}
\description{
Metadata about a CRAN package
}
\examples{
\dontshow{if (pingr::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cran_package("pkgsearch")
\dontshow{\}) # examplesIf}
}
