% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb-public-api.R
\name{cran_package_history}
\alias{cran_package_history}
\title{Query the history of a package}
\usage{
cran_package_history(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
A tibble, with one row per package version.
}
\description{
Query the history of a package
}
\examples{
\dontshow{ asNamespace("pkgsearch")$exif(pingr::is_online(), \{ }
cran_package_history("igraph")
\dontshow{ \}) }
}
