% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb-public-api.R
\name{cran_packages}
\alias{cran_packages}
\title{Metadata about multiple CRAN packages}
\usage{
cran_packages(names)
}
\arguments{
\item{names}{Package names. May also contain versions, separated by a
dash or \verb{@} character.}
}
\value{
A data frame of package metadata, one package per row.
}
\description{
Metadata about multiple CRAN packages
}
