% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb-public-api.R
\name{cran_releases}
\alias{cran_releases}
\title{List R releases in the CRANDB database}
\usage{
cran_releases()
}
\value{
List of R releases.
}
\description{
List R releases in the CRANDB database
}
\examples{
\dontrun{
cran_releases()
}

}
