% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb-public-api.R
\name{cran_top_downloaded}
\alias{cran_top_downloaded}
\title{Top downloaded packages}
\usage{
cran_top_downloaded()
}
\value{
Tibble of top downloaded packages.
}
\description{
Last week.
}
\details{
You can use the \href{https://r-hub.github.io/cranlogs/}{\code{cranlogs} package}
to get more flexibility into what is returned.
}
\examples{
\dontshow{ asNamespace("pkgsearch")$exif(FALSE, \{ }
cran_top_downloaded()
\dontshow{ \}) }
}
