% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb-public-api.R
\name{cran_trending}
\alias{cran_trending}
\title{Trending R packages}
\usage{
cran_trending()
}
\value{
Tibble of trending packages.
}
\description{
Trending packages are the ones that were downloaded at least 1000 times
during last week, and that substantially increased their download
counts, compared to the average weekly downloads in the previous 24
weeks. The percentage of increase is also shown in the output.
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cran_trending()
\dontshow{\}) # examplesIf}
}
