% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{list_cran_packages}
\alias{list_cran_packages}
\title{List all packages in a CRAN mirror.}
\usage{
list_cran_packages()
}
\description{
This includes archived packages, but currently does
not include packages whose name was reused by another package.
}
