% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb-public-api.R
\name{list_packages}
\alias{list_packages}
\title{List active packages}
\usage{
list_packages(from = "", limit = 10, format = c("short", "latest",
  "full"), archived = FALSE)
}
\arguments{
\item{from}{The name of the first package to list. By default it
is the first one in alphabetical order.}

\item{limit}{The number of packages to list.}

\item{format}{What to return. \sQuote{\code{short}} means the
title and version number only. \sQuote{\code{latest}} means
the complete description of the latest version. \sQuote{\code{full}}
means all versions.}

\item{archived}{Whether to include archived packages in the result.
If this is \code{TRUE}, then \code{format} must be
\sQuote{\code{full}}.}
}
\value{
List of packages.
}
\description{
List active packages
}
\examples{
\dontrun{
list_packages(from = "pkgsearch")
list_packages(format = "latest")
list_packages(format = "full")
}

}
