% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{more}
\alias{more}
\title{Next page after a seer search}
\usage{
more(format = c("short", "long"), size)
}
\arguments{
\item{format}{Default formatting of the results. \sQuote{short} only
outputs the name and title of the packages, \sQuote{long} also
prints the author, last version, full description and URLs.
Note that this only affects the default printing, and you can
still inspect the full results, even if you specify \sQuote{short}
here.}

\item{size}{The number of results to list.}
}
\description{
Next page after a seer search
}
\seealso{
Other seer: \code{\link{see}}
}
\concept{seer}
