% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pkg_search}
\alias{pkg_search}
\title{Search CRAN packages}
\usage{
pkg_search(query, format = c("short", "long"), from = 1, size = 10)
}
\arguments{
\item{query}{Search query string. If this argument is missing
then the results of the last query are printed, in \sQuote{short}
and \sQuote{long} formats, in turns for successive \code{pkg_search()}
calls. If this argument is missing, then all other arguments
are ignored.}

\item{format}{Default formatting of the results. \sQuote{short} only
outputs the name and title of the packages, \sQuote{long} also
prints the author, last version, full description and URLs.
Note that this only affects the default printing, and you can
still inspect the full results, even if you specify \sQuote{short}
here.}

\item{from}{Where to start listing the results, for pagination.}

\item{size}{The number of results to list.}
}
\value{
A list of packages, in a special object.
}
\description{
Search CRAN packages
}
\details{
Note that the search needs a working Internet connection.
}
\examples{
\donttest{
## Some example searches
pkg_search("networks")
pkg_search("survival")
pkg_search("graphics")
pkg_search("google")
}
}
