% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{pkg_search_addin}
\alias{pkg_search_addin}
\title{RStudio addin to search CRAN packages}
\usage{
pkg_search_addin(query = "", viewer = c("dialog", "browser"))
}
\arguments{
\item{query}{Query string to start the addin with.}

\item{viewer}{Whether to show the addin within RStudio (\code{"dialog"}),
or in a web browser (\code{"browser"}).}
}
\description{
Call this function from RStudio.
}
