% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_search.R
\name{pkg_search_adv}
\alias{pkg_search_adv}
\title{Advanced search of CRAN packages}
\usage{
pkg_search_adv(query = NULL, format = c("short", "long"), from = 1,
  size = 10)
}
\arguments{
\item{format}{Default formatting of the results. \emph{short} only
outputs the name and title of the packages, \emph{long} also
prints the author, last version, full description and URLs.
Note that this only affects the default printing, and you can
still inspect the full results, even if you specify \emph{short}
here.}

\item{from}{Where to start listing the results, for pagination.}

\item{size}{The number of results to list.}

\item{package}{Query on the package name}

\item{title}{Query on the package title}

\item{author}{Query on the package authors}

\item{maintainer}{Query on the package maintainer}

\item{description}{Query on the package description}

\item{keywords}{Query on the package keywords}
}
\value{
A tibble with columns:
\itemize{
\item \code{score}: Score of the hit. See Section \emph{Scoring} for some details.
\item \code{package}: Package name.
\item \code{version}: Latest package version.
\item \code{title}: Package title.
\item \code{description}: Short package description.
\item \code{date}: Time stamp of the last release.
\item \code{maintainer_name}: Name of the package maintainer.
\item \code{maintainer_email}: Email address of the package maintainer.
\item \code{revdeps}: Number of (strong and weak) reverse dependencies of the
package.
\item \code{downloads_last_month}: Raw number of package downloads last month,
from the RStudio CRAN mirror.
\item \code{license}: Package license.
\item \code{url}: Package URL(s).
\item \code{bugreports}: URL of issue tracker, or email address for bug reports.
}
}
\description{
\code{pkg_search_adv()} starts a new advanced search query.
}
\details{
Note that the search needs a working Internet connection.
}
