% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsearch-package.R
\docType{package}
\name{pkgsearch-package}
\alias{pkgsearch}
\alias{pkgsearch-package}
\title{pkgsearch: Search CRAN R Packages and get their metadata}
\description{
Search CRAN R packages and their metadata. 
  Uses the 'METACRAN' search server,
  see <https://r-pkg.org> and the CRAN metadata database, that 
  contains information about CRAN packages. Note that this is _not_ 
  a CRAN project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-hub/pkgsearch}
  \item \url{https://r-hub.github.io/pkgsearch}
  \item Report bugs at \url{https://github.com/r-hub/pkgsearch/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item R Consortium [funder]
}

}
\keyword{internal}
