% Generated by roxygen2 (4.0.1): do not edit by hand
\name{seer}
\alias{seer}
\title{Search CRAN packages}
\usage{
seer(query, from = 0, size = 20, server = "seer.r-pkg.org", port = 9200)
}
\arguments{
\item{query}{Search query string.}

\item{from}{Where to start listing the results, for pagination.}

\item{size}{The number of results to list.}

\item{server}{The server to use.}

\item{port}{Port to use.}
}
\value{
A list of packages, in a special object.
}
\description{
Search CRAN packages
}
\details{
Note that the search needs a working Internet connection.
}
\examples{
\donttest{
## Some example searches
seer("networks")
seer("survival")
seer("graphics")
seer("google")
}
}

