% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-package-list.R
\name{cran-package-list}
\alias{cran_package_list}
\title{Lists of CRAN packages}
\usage{
cran_package_list(from = "", limit = 10, archived = FALSE)
}
\arguments{
\item{from}{The name of the first package to list. By default it
is the first one in alphabetical order.}

\item{limit}{The number of packages to list.}

\item{archived}{Whether to include archived packages in the result.
If this is \code{TRUE}, then \code{format} must be
\sQuote{\code{full}}.}
}
\value{
A \code{tbl_df} with two columns:
\itemize{
\item version
\item title
}
}
\description{
List packages on CRAN, in different formats.
}
\section{\code{cran_package_list()}}{

For all packages, returns title and version number only.
}

