% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crandb-public-api.R
\name{cran_events}
\alias{cran_events}
\title{List of all CRAN events (new, updated, archived packages)}
\usage{
cran_events(limit = 10, releases = TRUE, archivals = TRUE)
}
\arguments{
\item{limit}{Number of events to list.}

\item{releases}{Whether to include package releases.}

\item{archivals}{Whether to include package archivals.}
}
\value{
List of events.
}
\description{
List of all CRAN events (new, updated, archived packages)
}
