% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{pkg_search_addin}
\alias{pkg_search_addin}
\title{RStudio addin to search CRAN packages}
\usage{
pkg_search_addin(query = "", viewer = c("dialog", "browser"))
}
\arguments{
\item{query}{Query string to start the addin with.}

\item{viewer}{Whether to show the addin within RStudio (\code{"dialog"}),
or in a web browser (\code{"browser"}).}
}
\description{
Call this function from RStudio for best results. You can also use it
without RStudio, then it will run in the web browser.
}
\details{
The app has:
\itemize{
\item A search tab for free text search, very much like the \code{\link[=pkg_search]{pkg_search()}}
function.
\item The list of recently updated packages.
\item The list of top packages: most downloaded, most depended upon,
and trending packages.
\item Package list by maintainer.
}
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{pkg_search_addin()

# Start with a search query
pkg_search_addin("permutation test")
}\if{html}{\out{</div>}}
}

