% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsearch-package.R
\docType{package}
\name{pkgsearch-package}
\alias{pkgsearch}
\alias{pkgsearch-package}
\title{pkgsearch: Search and Query CRAN R Packages}
\description{
Search CRAN metadata about packages by keyword, popularity, recent activity, package name and more. Uses the 'R-hub' search server, see <https://r-pkg.org> and the CRAN metadata database, that contains information about CRAN packages. Note that this is _not_ a CRAN project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-hub/pkgsearch}
  \item \url{https://r-hub.github.io/pkgsearch/}
  \item Report bugs at \url{https://github.com/r-hub/pkgsearch/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Maëlle Salmon (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
}

Other contributors:
\itemize{
  \item R Consortium [funder]
}

}
\keyword{internal}
